+++ Cargo.toml
@@ -18,6 +18,7 @@ octocrab = { version = "0.8.13", default-features = false }
 tokio = { version = "1.0", default-features = false, features = ["macros", "rt-multi-thread", "time"] }
 # graphql_client = "0.9"
 serde = { version = "1.0.106", features = ["derive"] }
+serde_urlencoded = "0.7.0"
 reqwest = { version = "0.11.0", default-features = false, features = ["rustls-tls", "json"] }
 #prettytable-rs = "^0.7"
 structopt = "^0.3"
